# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bulls and Cows game",
        "description": "Game API",
        "version": "0.0.1"
    },
    "paths": {
        "/api/game/create": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Game create",
                "operationId": "ac0c37c02739b99cedc0f4fabc868823",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "username",
                                    "user_age"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "User name",
                                        "type": "string"
                                    },
                                    "user_age": {
                                        "description": "User age",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game created"
                    },
                    "400": {
                        "description": "Input erros"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/game/delete/{id}": {
            "delete": {
                "tags": [
                    "Game"
                ],
                "summary": "Delete a game",
                "operationId": "f104f44523cfaba739c8966756d3be5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Enter a game id",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game delete successfully"
                    },
                    "404": {
                        "description": "The game has been not found"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/game/proposeCombination": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Analyze a combination",
                "operationId": "18893f84144f69967f82211c05d3aa21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "combination",
                                    "id"
                                ],
                                "properties": {
                                    "combination": {
                                        "description": "combination",
                                        "type": "string"
                                    },
                                    "id": {
                                        "description": "Game id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game info"
                    },
                    "403": {
                        "description": "Duplicate value in combination or duplicate combination"
                    },
                    "404": {
                        "description": "Game not found"
                    },
                    "408": {
                        "description": "Game over"
                    },
                    "409": {
                        "description": "Unavailable game, it was won"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/game/previewResponse": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Get a preview response",
                "operationId": "eaa60b20952ea4429705d767ca14caaa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "attempt",
                                    "id"
                                ],
                                "properties": {
                                    "attempt": {
                                        "description": "attempt number",
                                        "type": "integer"
                                    },
                                    "id": {
                                        "description": "Game id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game info"
                    },
                    "403": {
                        "description": "There is no information for attempt"
                    },
                    "404": {
                        "description": "Game not found"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            }
        }
    }
}